+++ src/attestation/mod.rs
@@ -4,4 +4,7 @@
 
 pub mod evidence;
 
+#[cfg(any(feature = "openssl", feature = "crypto_nossl"))]
+pub mod verifier;
+
 pub use evidence::snp::{KeyInfo, PlatformInfo, Report, ReportBody, ReportVariant};